# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для безопасного хранения паролей, текстов, карт и файлов.",
        "title": "GophKeeper API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/storage/auth-info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает сохранённые логин/пароль по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "Получить учётные данные",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID записи",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_konkovaanna23_gophkeeper_pkg_keeperservice.StoredAuthInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет запись с проверкой версии",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "Обновить учётные данные",
                "parameters": [
                    {
                        "description": "ID, data, version",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_konkovaanna23_gophkeeper_pkg_keeperservice.UpdateAuthInfoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Конфликт версий",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Сохраняет логин и пароль",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "Сохранить учётные данные",
                "parameters": [
                    {
                        "description": "Данные: login, password, meta",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_konkovaanna23_gophkeeper_pkg_keeperservice.AuthInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет запись по ID и версии",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "Удалить учётные данные",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID записи",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Текущая версия",
                        "name": "version",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_konkovaanna23_gophkeeper_pkg_keeperservice.DeleteInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.CreateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "rec-123"
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "пользователь не найден"
                }
            }
        },
...
        "dto.UpdateResponse": {
            "type": "object",
            "properties": {
                "new_version": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "github_com_konkovaanna23_gophkeeper_pkg_keeperservice.AuthInfo": {
            "type": "object",
            "properties": {
                "login": {
                    "description": "login — логин (например, email).",
                    "type": "string"
                },
                "meta": {
                    "description": "meta — произвольные метаданные (например, \"название\", \"источник\").",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "password": {
                    "description": "password — пароль (зашифрован на клиенте или сервере).",
                    "type": "string"
                }
            }
        },
...
        "github_com_konkovaanna23_gophkeeper_pkg_keeperservice.DeleteInfoResponse": {
            "type": "object"
        },
        "github_com_konkovaanna23_gophkeeper_pkg_keeperservice.StoredAuthInfo": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "data — содержимое записи.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_konkovaanna23_gophkeeper_pkg_keeperservice.AuthInfo"
                        }
                    ]
                },
                "id": {
                    "description": "id — уникальный идентификатор записи.",
                    "type": "string"
                },
                "updated_at": {
                    "description": "updated_at — время последнего обновления.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/timestamppb.Timestamp"
                        }
                    ]
                },
                "version": {
                    "description": "version — номер версии (для контроля конфликтов при обновлении).",
                    "type": "integer"
                }
            }
        },
...
        "github_com_konkovaanna23_gophkeeper_pkg_keeperservice.UpdateAuthInfoRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_konkovaanna23_gophkeeper_pkg_keeperservice.AuthInfo"
                },
                "id": {
                    "type": "string"
                },
                "version": {
                    "description": "version — текущая версия записи (защита от конфликтов).",
                    "type": "integer"
                }
            }
        },
...
        "timestamppb.Timestamp": {
            "type": "object",
            "properties": {
                "nanos": {
                    "description": "Non-negative fractions of a second at nanosecond resolution. Negative\nsecond values with fractions must still have non-negative nanos values\nthat count forward in time. Must be from 0 to 999,999,999\ninclusive.",
                    "type": "integer"
                },
                "seconds": {
                    "description": "Represents seconds of UTC time since Unix epoch\n1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to\n9999-12-31T23:59:59Z inclusive.",
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer Token: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}